backend/requirements.txt
alembic==1.14.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
asyncpg==0.30.0
babel==2.16.0
bcrypt==4.0.1
black==24.10.0
blobfile==3.0.0
capstone==6.0.0a1
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
colored-traceback==0.4.2
comm==0.2.2
contourpy==1.3.0
coverage==7.6.2
cryptography==43.0.3
cycler==0.12.1
databases==0.9.0
debugpy==1.8.8
decorator==5.1.1
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.6
filelock==3.16.1
fire==0.7.0
flake8==7.1.1
fonttools==4.54.1
fsspec==2024.9.0
gmpy2==2.2.1
graphql-core==3.2.5
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
identify==2.6.1
idna==3.10
iniconfig==2.0.0
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
llama_models==0.0.42
llama_stack==0.0.42
lxml==5.3.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
mypy-extensions==1.0.0
Naked==0.1.32
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
paramiko==3.5.0
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
plumbum==1.9.0
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pwntools==4.13.1
pyasn1==0.6.1
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic-settings==2.7.0
pydantic_core==2.23.4
pyelftools==0.31
pyflakes==3.2.0
Pygments==2.18.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyNaCl==1.5.0
pyparsing==3.2.0
pyserial==3.5
PySocks==1.7.1
pytest==8.3.3
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2024.2
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
rich==13.9.2
ROPGadget==7.5
rpyc==6.0.1
rsa==4.9
setuptools==75.1.0
shellescape==3.8.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
strawberry-graphql==0.255.0
sympy==1.13.3
termcolor==2.5.0
tiktoken==0.8.0
tkcalendar==1.6.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
unicorn==2.1.1
unix-ar==0.2.1
urllib3==2.2.3
uvicorn==0.34.0
virtualenv==20.26.6
wcwidth==0.2.13
zstandard==0.23.0
